Fix method redefinition warnings when pluralizations are reloaded
Fix: https://github.com/svenfuchs/rails-i18n/issues/735

Whenever pluralization rules are reloaded it causes a lot of warnings:

```
lib/rails_i18n/../../rails/pluralization/ar.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/ar.rb:4: warning: previous definition of rule was here
lib/rails_i18n/../../rails/pluralization/gd.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/gd.rb:4: warning: previous definition of rule was here
lib/rails_i18n/../../rails/pluralization/hsb.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/hsb.rb:4: warning: previous definition of rule was here
lib/rails_i18n/../../rails/pluralization/lt.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/lt.rb:4: warning: previous definition of rule was here
lib/rails_i18n/../../rails/pluralization/lv.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/lv.rb:4: warning: previous definition of rule was here
lib/rails_i18n/../../rails/pluralization/mk.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/mk.rb:4: warning: previous definition of rule was here
lib/rails_i18n/../../rails/pluralization/pl.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/pl.rb:4: warning: previous definition of rule was here
lib/rails_i18n/../../rails/pluralization/sl.rb:4: warning: method redefined; discarding old rule
lib/rails_i18n/../../rails/pluralization/sl.rb:4: warning: previous definition of rule was here
...
```

This is mostly harmless, but annoying for users who try to run with no warnings at all.
