+++ rails/pluralization/mk.rb
@@ -1,18 +1,4 @@
-module RailsI18n
-  module Pluralization
-    module Macedonian
-      def self.rule
-        lambda do |n|
-          if n.is_a?(Numeric) && n % 10 == 1 && n != 11
-            :one
-          else
-            :other
-          end
-        end
-      end
-    end
-  end
-end
+require 'rails_i18n/pluralization'
 
 { :mk => {
     :'i18n' => {
