+++ rails/pluralization/sl.rb
@@ -1,21 +1,4 @@
-module RailsI18n
-  module Pluralization
-    module Slovenian
-      def self.rule
-        lambda do |n|
-          return :other unless n.is_a?(Numeric)
-
-          case n % 100
-          when 1 then :one
-          when 2 then :two
-          when 3, 4 then :few
-          else :other
-          end
-        end
-      end
-    end
-  end
-end
+require 'rails_i18n/pluralization'
 
 { :sl => {
     :'i18n' => {
